chore: rename Web3Names folder to DidLinks
